move lib/index to cli.js; - no longer need harmony flag
